refactor: improve names and factories
