+++ scafi-distributed/js/src/test/scala/it/unibo/scafi/runtime/network/sockets/JSSocketNetworkingTest.scala
@@ -2,7 +2,7 @@ package it.unibo.scafi.runtime.network.sockets
 
 import scala.concurrent.ExecutionContext
 
-class JSNativeSocketNetworkingTest extends SocketNetworkingBehavior:
+class JSSocketNetworkingTest extends SocketNetworkingBehavior:
 
   override given executionContext: ExecutionContext = scalajs.concurrent.JSExecutionContext.Implicits.queue
