+++ scafi-distributed/jvm-native/src/main/scala/it/unibo/scafi/runtime/network/sockets/SocketNetworking.scala
@@ -7,7 +7,7 @@ import scala.LazyList.continually
 import scala.concurrent.{ ExecutionContext, Future }
 import scala.util.{ Failure, Success, Try }
 
-trait SocketNetworking(using ec: ExecutionContext, conf: SocketConfiguration) extends ConnectionOrientedTemplate:
+trait SocketNetworking(using ec: ExecutionContext, conf: ConnectionConfiguration) extends ConnectionOrientedTemplate:
 
   override def out(endpoint: Endpoint): Future[Connection] =
     for
