+++ scafi-distributed/shared/src/main/scala/it/unibo/scafi/runtime/network/sockets/ConnectionConfiguration.scala
@@ -3,9 +3,9 @@ package it.unibo.scafi.runtime.network.sockets
 import scala.concurrent.duration.{ DurationInt, FiniteDuration }
 
 /**
- * Socket-based networking configuration.
+ * Connection oriented networking configuration.
  */
-trait SocketConfiguration:
+trait ConnectionConfiguration:
 
   /** The maximum acceptable size of a message in bytes that can be received over the socket stream. */
   val maxMessageSize: Int
@@ -13,12 +13,12 @@ trait SocketConfiguration:
   /** The maximum duration that a connection may remain open without receiving any messages. */
   val inactivityTimeout: FiniteDuration
 
-object SocketConfiguration:
+object ConnectionConfiguration:
 
   /**
-   * A basic socket configuration with a maximum message size of 65,536 bytes and an inactivity timeout of 60 seconds.
+   * A basic configuration set up with a maximum message size of 65,536 bytes and an inactivity timeout of 60 seconds.
    */
-  def basic: SocketConfiguration = new SocketConfiguration:
+  def basic: ConnectionConfiguration = new ConnectionConfiguration:
     override val maxMessageSize: Int = 65_536
     override val inactivityTimeout: FiniteDuration = 60.seconds
