+++ scafi-distributed/shared/src/test/scala/it/unibo/scafi/runtime/network/sockets/SocketNetworkingBehavior.scala
@@ -12,7 +12,7 @@ import org.scalatest.time.{ Millis, Seconds, Span }
 
 trait SocketNetworkingBehavior extends AsyncSpec:
 
-  given SocketConfiguration = SocketConfiguration.basic
+  given ConnectionConfiguration = ConnectionConfiguration.basic
 
   given PatienceConfig = PatienceConfig(timeout = Span(2, Seconds), interval = Span(250, Millis))
 
@@ -41,7 +41,7 @@ trait SocketNetworkingBehavior extends AsyncSpec:
         _ shouldBe a[Throwable]
 
     it should "be able to connect to an available remote endpoint" in:
-      usingServer(nop): (_, client) =>
+      usingServer(nop): client =>
         client.isOpen shouldBe true
         client.close()
         client.isOpen shouldBe false
@@ -50,21 +50,21 @@ trait SocketNetworkingBehavior extends AsyncSpec:
     it should "be able to accept incoming connections from remote endpoints" in:
       val receivedMessages = CopyOnWriteArrayList[net.MessageIn]()
       val messages = Seq("Hello", "World", "Scafi", "Networking")
-      usingServer(msg => receivedMessages.add(msg): Unit): (_, client) =>
+      usingServer(msg => receivedMessages.add(msg): Unit): client =>
         client send messages verifying eventually(receivedMessages should contain theSameElementsInOrderAs messages)
 
     it should "close connections with clients attempting to flood the server" in:
       val tooLargeMessage = "A" * 65_536
-      usingServer(nop): (_, client) =>
+      usingServer(nop): client =>
         client send Seq(tooLargeMessage) verifying eventually(client.isOpen shouldBe false)
 
   private def usingServer[Result](using
       net: ConnectionOrientedNetworking,
-  )(onMessage: net.MessageIn => Unit)(todo: (net.ListenerRef, net.Connection) => Future[Result]) =
+  )(onMessage: net.MessageIn => Unit)(todo: net.Connection => Future[Result]) =
     for
       server <- net.in(FreePort)(onMessage)
       client <- net.out(Endpoint(Localhost, server.listener.boundPort))
-      result <- todo(server, client)
+      result <- todo(client)
       _ = client.close()
       _ = server.listener.close()
     yield result
